Solution to Check Not Equal To

+++ Check Not Equal To@@67882a731b4392d8fc1577db/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    double num1, num2;
+    int num1, num2;
     
-    scanf("%lf", &num1);
+    scanf("%d", &num1);
     
 
-    scanf("%lf", &num2);
+    scanf("%d", &num2);
     
     // Checking if num1 is equal to num2
     if (num1 != num2) {
